+++ src/test/sample/case/case1.lua
@@ -1,0 +1,8 @@
+local Case1 = {}
+
+local function local_func_export()
+end
+
+Case1.local_func_export = local_func_export
+
+return Case1