+++ src/test/sample/case/test1.lua
@@ -1,0 +1,2 @@
+local case1 = require "case1"
+case1.local_func_export()